# backend/src/main/java/com/example/backend/dtos/OwnAccountResponse.java
package com.example.backend.dtos;

import com.example.backend.entities.enums.AccountStatus;
import com.example.backend.entities.enums.AccountType;

import java.math.BigDecimal;
import java.time.LocalDateTime;

public record OwnAccountResponse(
        String iban,
        AccountType type,
        BigDecimal balance,
        BigDecimal absoluteTransferLimit,
        BigDecimal dailyTransferLimit,
        AccountStatus status,
        LocalDateTime createdAt
) {}