# backend/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: ATM Booth Banking API
  version: 0.0.1
  description: REST backend for the ATM Booth full-stack project (InHolland). Demo users are seeded on startup — see README.
servers:
  - url: http://localhost:8080
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registered user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error or duplicate email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and receive an access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags: [Auth]
      summary: Get the currently authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /accounts:
    get:
      tags: [Accounts]
      summary: Employee account search (paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: integer
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AccountType'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AccountStatus'
        - name: iban
          in: query
          required: false
          schema:
            type: string
        - name: name
          in: query
          required: false
          description: Search by customer first name, last name or email
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
        - name: sort
          in: query
          required: false
          description: Sort expression such as `createdAt,desc`
          schema:
            type: string
      responses:
        '200':
          description: Page of private employee account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeAccountPage'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /accounts/me:
    get:
      tags: [Accounts]
      summary: List the current customer's accounts
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
        - name: sort
          in: query
          required: false
          description: Sort expression such as `createdAt,desc`
          schema:
            type: string
      responses:
        '200':
          description: Page of the current active customer's accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnAccountPage'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied or customer is not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /accounts/transfer-targets:
    get:
      tags: [Accounts]
      summary: Search public transfer targets
      description: Returns active checking accounts owned by other customers. Response rows expose only IBAN and customer name fields.
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: true
          description: Customer first or last name fragment
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
        - name: sort
          in: query
          required: false
          description: Sort expression such as `iban,asc`
          schema:
            type: string
      responses:
        '200':
          description: Page of safe transfer-target results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferTargetPage'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied or customer is not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /accounts/{iban}:
    patch:
      tags: [Accounts]
      summary: Update account limits or status
      description: At least one field must be provided. Setting status to CLOSED requires a zero balance.
      security:
        - bearerAuth: []
      parameters:
        - name: iban
          in: path
          required: true
          schema:
            type: string
            example: NL91ABNA0417164300
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdateRequest'
      responses:
        '200':
          description: Updated account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeAccountResponse'
        '400':
          description: Validation error, no fields provided, or cannot close account with non-zero balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /transactions:
    get:
      tags: [Transactions]
      summary: List all transactions
      security:
        - bearerAuth: []
      parameters:
        - name: iban
          in: query
          required: false
          schema:
            type: string
        - name: customerId
          in: query
          required: false
          description: Filter by initiating customer. Ignored for authenticated customers; employees can filter any customer.
          schema:
            type: integer
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TransactionType'
        - name: minAmount
          in: query
          required: false
          schema:
            type: number
        - name: maxAmount
          in: query
          required: false
          schema:
            type: number
        - name: startDate
          in: query
          required: false
          description: Include transactions on or after this date.
          schema:
            type: string
            format: date
            example: '2026-05-01'
        - name: endDate
          in: query
          required: false
          description: Include transactions on or before this date.
          schema:
            type: string
            format: date
            example: '2026-05-31'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
        - name: sort
          in: query
          required: false
          description: Sort expression such as `timestamp,desc`
          schema:
            type: string
      responses:
        '200':
          description: Page of transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionPage'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Transactions]
      summary: Create a new transaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreateRequest'
      responses:
        '201':
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Validation error, missing required IBAN for transaction type, inactive account, insufficient transfer limit, or malformed request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Source or destination account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /transactions/{id}:
    get:
      tags: [Transactions]
      summary: Get transaction by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AccountType:
      type: string
      enum: [CHECKING, SAVINGS]
    AccountStatus:
      type: string
      enum: [ACTIVE, CLOSED]
    CustomerStatus:
      type: string
      enum: [PENDING, ACTIVE, CLOSED]
    TransactionType:
      type: string
      enum: [TRANSFER, DEPOSIT, WITHDRAWAL]
    UserRole:
      type: string
      enum: [CUSTOMER, EMPLOYEE]

    # ── Requests ─────────────────────────────────────────────────────────────
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName, bsn, phoneNumber]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          pattern: '^(?=.*[A-Z])(?=.*[a-z])(?=.*[^a-zA-Z0-9]).{8,}$'
          minLength: 8
        firstName:
          type: string
          maxLength: 64
        lastName:
          type: string
          maxLength: 64
        bsn:
          type: string
          pattern: '^\d{8,9}$'
        phoneNumber:
          type: string
          pattern: '^06[0-9]{8}$'
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AccountUpdateRequest:
      type: object
      description: At least one field must be provided.
      properties:
        absoluteTransferLimit:
          type: number
          minimum: 0
        dailyTransferLimit:
          type: number
          minimum: 0
        status:
          $ref: '#/components/schemas/AccountStatus'
...
    TransactionCreateRequest:
      type: object
      description: TRANSFER requires both `fromIban` and `toIban`; DEPOSIT requires `toIban`; WITHDRAWAL requires `fromIban`.
      required: [amount, type]
      properties:
        fromIban:
          type: string
        toIban:
          type: string
        initiatedByUserId:
          type: integer
          deprecated: true
          description: Ignored for authenticated customer requests. The server uses the JWT principal as the authoritative initiator.
        amount:
          type: number
          exclusiveMinimum: true
          minimum: 0
        type:
          $ref: '#/components/schemas/TransactionType'
        description:
          type: string
          maxLength: 255

    # ── Responses ────────────────────────────────────────────────────────────
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        createdAt:
          type: string
          format: date-time
    LoginResponse:
      type: object
      properties:
        token:
          $ref: '#/components/schemas/TokenResponse'
        user:
          $ref: '#/components/schemas/CurrentUserResponse'
    TokenResponse:
      type: object
      properties:
        value:
          type: string
        expiration:
          type: integer
          format: int64
          description: Expiration time in seconds
        type:
          type: string
          example: Bearer
    CurrentUserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/CustomerStatus'
        bsn:
          type: string
        phoneNumber:
          type: string
    EmployeeAccountResponse:
      type: object
      properties:
        userId:
          type: integer
        iban:
          type: string
        type:
          $ref: '#/components/schemas/AccountType'
        balance:
          type: number
        absoluteTransferLimit:
          type: number
        dailyTransferLimit:
          type: number
        status:
          $ref: '#/components/schemas/AccountStatus'
        createdAt:
          type: string
          format: date-time
    OwnAccountResponse:
      type: object
      properties:
        iban:
          type: string
        type:
          $ref: '#/components/schemas/AccountType'
        balance:
          type: number
        absoluteTransferLimit:
          type: number
        dailyTransferLimit:
          type: number
        status:
          $ref: '#/components/schemas/AccountStatus'
        createdAt:
          type: string
          format: date-time
...
    TransferTargetResponse:
      type: object
      description: Safe customer-facing transfer target result. Does not expose balances or limits.
      properties:
        iban:
          type: string
        firstName:
          type: string
        lastName:
          type: string
...
    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
        fromIban:
          type: string
        toIban:
          type: string
        initiatedByUserId:
          type: integer
        amount:
          type: number
        type:
          $ref: '#/components/schemas/TransactionType'
        description:
          type: string
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: string
    EmployeeAccountPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/EmployeeAccountResponse'
    OwnAccountPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/OwnAccountResponse'
    TransferTargetPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/TransferTargetResponse'
...
    TransactionPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/TransactionResponse'
    PageMetadata:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        pageable:
          $ref: '#/components/schemas/PageableMetadata'
        sort:
          $ref: '#/components/schemas/SortMetadata'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
        empty:
          type: boolean
    PageableMetadata:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        offset:
          type: integer
          format: int64
        paged:
          type: boolean
        unpaged:
          type: boolean
        sort:
          $ref: '#/components/schemas/SortMetadata'
    SortMetadata:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean